requirements.txt
fastapi==0.95.1
locust==2.16.1
loguru==0.7.0
numpy==1.24.3
pandas==1.4.4
pydantic==1.10.7
pytest==7.4.0
python-multipart==0.0.6
PyYAML==6.0
scikit-learn==1.1.2
uvicorn==0.22.0
